Moved code handling swagger-ui into lib/swagger-ui.js
